src/meet-bot/requirements.txt
google-api-python-client
google-auth-oauthlib
playwright
requests
